Fix spurious unhandled rejection reporting.
Normally a client performs a call by sending a "push" followed by a "pull". But the client does not have to send a "pull" (especially when doing pipelining), and even if it does send a "pull", it will often not arrive in exactly the same tick as the "push".

If the called method throws immediately, though, this could be seen as an "unhandled rejection", and reported as such on the callee's side. It's common for applications to crash on unhandled rejections. But in this case the rejection is the caller's responsibility, so the callee crashing is rather bad.

To solve this, on a "pull", we must mark all promises in the response payload as having had errors handled, using a no-op `catch`.

The tests missed this bug previously because the test transport does not do any I/O, and thus all messages are seen as delivered immediately, so the promises don't get a chance to be considered unhandled.