Change stub ownership transfer rules in call params.
It's been wrong all along. Unfortunately it's wrong in the original workerd built-in implementation as well.

Specifically, historically, the rule has been: When making a call, the caller loses ownership of any stubs passed in the params. If you didn't want to transfer ownership away, you were supposed to `.dup()` the stub.

But the introduction of promise pipelining made this obviously ugly. If this rule applies to promises in addition to stubs, then you end up with code like:

```
let namePromise = api.whoami();
let greeting = await api.greet(namePromise.dup());

let name = await namePromise;
```

This need to `.dup()` everything in promise pipelining did not sit well with me, so I tried to solve it months ago, but I solved it in the wrong way. I decided to say that we never implicitly transfer ownership of promises, only stubs.

This made promise pipelining look nice! And it had the benefit that it didn't make any backwards-incompatible changes in the semantics compared to what's in workerd (since workerd's RPC doesn't actually support embedding promises in payloads at all, yet).

But:
* This applied to return values too. And that basically meant you can't return a promise, because how would you ever dispose it? That kind of sucks, because returning a promise is useful when proxying.
* Even in the absence of promise pipelining, the original behavior is problematic for proxying too: when proxying a call, you need to manually dup() all the stubs in the params.

Ugh!

The right answer is: Don't implicitly transfer ownership of stubs in params. ONLY transfer ownership of stubs in results (including promises).

This is a change from workerd, but probably one we should backport, as it makes proxying easier.